+++ crates/fluvio-cli/src/consume/record_format.rs
@@ -73,13 +73,10 @@ pub fn format_raw_record(record: &[u8]) -> String {
 //  Table
 // -----------------------------------
 
-// TODO: This will eventually read for alternative formatting
 /// Print records in table format
 pub fn print_table_record(record: &[u8], count: i32) -> String {
     use prettytable::{Row, cell, Cell, Slice};
     use prettytable::format::{self, FormatBuilder};
-
-    //let mut hashmap : std::collections::HashMap<String, serde_json::Value> = std::collections::HashMap::new();
 
     let maybe_json: serde_json::Value = match serde_json::from_slice(record) {
         Ok(value) => value,
@@ -107,23 +104,11 @@ pub fn print_table_record(record: &[u8], count: i32) -> String {
         .collect();
 
     let header: Row = Row::new(keys_str.iter().map(|k| cell!(k.to_owned())).collect());
-    let entries: Row = Row::new(
-        values_str
-            .iter()
-            .map(|v| {
-                let c = Cell::new(v);
-                c
-            })
-            .collect(),
-    );
+    let entries: Row = Row::new(values_str.iter().map(|v| Cell::new(v)).collect());
 
     // Print the table
-    let mut t_print = Vec::new();
-    // TODO: Don't print the headers in this table display
+    let t_print = vec![header, entries];
 
-    t_print.push(header);
-
-    t_print.push(entries);
     let mut table = prettytable::Table::init(t_print);
 
     let base_format: FormatBuilder = (*format::consts::FORMAT_NO_BORDER_LINE_SEPARATOR).into();
@@ -131,6 +116,7 @@ pub fn print_table_record(record: &[u8], count: i32) -> String {
     table.set_format(table_format.build());
 
     // FIXME: Live display of table data easily misaligns column widths
+    // if there is a length diff between the header and the data
     // The rows after the first (count == 0) don't line up with the header
     // prettytable might not support the live display use-case we want
     if count == 0 {