+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Add `#[smarstream(array_map)]` for expanding one record into many ([#1335](https://github.com/infinyon/fluvio/issues/1335))
 * Add capability to use input parameters in smartstreams ([#1643](https://github.com/infinyon/fluvio/issues/1643))
 * Make it easier to debug inline chart ([#1779](https://github.com/infinyon/fluvio/issues/1779))
+* Add table display output option to consumer for json objects ([#1642](https://github.com/infinyon/fluvio/issues/1642))
 
 ## Platform Version 0.9.10 - 2021-10-07
 * Improve error handling for socket timeout ([#791](https://github.com/infinyon/fluvio/issues/791))